+++ lab_3/main.py
@@ -1,16 +1,15 @@
 ﻿from src.cast5 import CAST5
+from src.rsa import RSA
 
 
 def main():
-    cipher = CAST5(40)
+    cipher = RSA()
+    
+    data = b"hello man"
+    c_d = cipher.encrypt(data)
+    d_d = cipher.decrypt(c_d)
 
-    data = b"hello world"
-
-    e_d = cipher.encrypt(data)
-    d_d = cipher.decrypt(e_d)
-
-    print(e_d, "\n", d_d, sep="")
-
+    print(c_d, "\n", d_d, sep="")
 
 if __name__ == "__main__":
     main()
