--- lab_3/src/hybrid.py
@@ -1,2 +1,0 @@
-﻿class HybridCryptoSystem:
-    pass
